+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Enhancement) Add support for `fmc_device_vti_interface`
 - (Enhancement) Add support for IKEv1 & IKEv2 IPSec Proposals & Policies
 - (Enhancement) Honor proxy settings (`HTTP_PROXY`, `HTTPS_PROXY`, `NO_PROXY` environment variables)
+- (Enhancement) Add support for Site-to-Site VPNs (`fmc_vpn_s2s`, `fmc_vpn_s2s_ipsec_settings`, `fmc_vpn_s2s_ike_settings`, `fmc_vpn_s2s_advanced_settings`, `fmc_vpn_s2s_endpoints`)
 
 ## 2.0.0-beta3